Moves project-specific bitters adds out of the bitters dir
We want to be able to plunk in a fresh styling directory with `bitters
install`, so any customizations within the bitters directory itself
will interfere with that. This commit moves these into `main` instead,
and adjusts any `@imports` such that the files in `bitters-seattle`
better reflect the file structure of a fresh Bitters install.

Renames _basic_grid -> _scorecard_grid since this only seems to be
referenced from the scorecard.