Allows setting the city theme via gulp
We'll want to switch the theme at compile time, as part of the gulp
packaging pipeline. The simplest way to do this is to set a global
argument to gulp, that can be made available to tasks:

`gulp <taskname> --city springfield`

If no argument is supplied, gulp defaults safely to `seattle.`

This commit adds `gulp-mustache` to allow generating of the `main.scss`
file. We need to do a filesystem-level switch instead of some logic
within the SASS file since it seems like dynamic dependencies in SASS
might be some ways off, see https://github.com/sass/sass/issues/739
(this would allow something like `@use "bitters-#{$city}/base"`, but
this doesn't seem like it's going to happen any time soon).

Beyond that, it will probably be worth switching other things on the
build pipeline rather than in code, eg. picking which config json from
`src/cities` ends up in the distribution, so this is a first step in
this direction.