Sets up SASS theming colors
Turns out `#0047ba` is not actually orange. Abstracts out Seattle's blue and
dark blue as semantically the "theme color" and the "theme color dark"
outside of the variables file so these can be more generally understood.
Removes `$normal-text` since this was only used in one place and
replaced it with the `$base-font-color` (semantically the closest
match). Sets the springfield theme to a bright red to make it simple to
spot anywhere that's been missed.